TIM5 舵机驱动重构 + P2 建系修复 + 相机-吸嘴偏置校准 + FPU 启用 + PnP 流程改进
TIM5 舵机驱动：
- TIM5 Prescaler 169→0, Period 19999→3399999（移除分频，直接 170MHz 计数）
- driver_servo.c/h: uint16_t→uint32_t 适配大 Period 值
- 新增 TIM5_IRQHandler 中断处理 + NVIC 配置

系统级改进：
- FreeRTOSConfig.h: configENABLE_FPU 0→1 启用硬件浮点
- app_freertos.c: 恢复 Host_Task 启动（替代 CamTestTask）
  ESP_Task 恢复启用（替代 StartESPTestTask）
  Key_Task 栈 256→512

P2 Mark 建系修复：
- 超时改用 tick（g_busy_enter_tick）替代累加计数器，避免不同主循环速率漂移
- 新增 P2_MAX_ALIGN_ITER(5) 防止对齐迭代死循环
- Mark 偏移修正使用 CAM_PX_TO_STEPS 宏统一换算
- g_marks_actual 记录真实机器坐标（吸嘴=摄像头+cam_to_nozzle）

相机-吸嘴偏置校准：
- 全局 cam_to_nozzle 默认值更新：(-5529,-25041)→(-6655,-24857)
- 所有散料区移动加 cam_offset 补偿（P1 扫描/P2 跳转/吸取重试）
- move_to_pcb_step 移除冗余偏移（g_marks_actual 已是真实坐标）

PnP 流程改进：
- PnP 散料区移动加 cam_offset 补偿（摄像头→吸嘴）
- PICK_DELAY_MS 300→800 等待 R 轴停稳
- host_correct_r_from_vision: 矫正前先 r_axis_set_zero() 重新校准零点
- offset_check_step: 所有出口加 LowerCam_Light_Off + Vision_SendEnd
- place_component: 吹气辅助脱料（Valve_On→800ms→Valve_Off）
- move_to_pcb_step: R 轴角度取反（r_axis_rotate(-target_angle)）
- Host_Task 启动加 P0 视觉握手（120s 超时）

运动控制：
- move_set_pad_ms(): 离散移动时临时 80ms 安全余量加速到位判断
- 电机错误统一进 HOST_ERROR（TIMEOUT 不再原地重试）
  + Vision_SendEnd + Vision_ForceIdle + Bridge_NotifyLog

视觉超时：
- VISION_TIMEOUT 30s→120s（P2 扫描+3 个 Mark 建系需要更长时间）
- P3 end 帧兼容 Phase1（p3_sub == P3_PHASE1）不再只匹配 Phase2

+++ Drivers/ZeMCU-G4/driver_servo.c
@@ -16,7 +16,7 @@ extern void PrintDebug(const char* fmt, ...);
  * ================================================================ */
 
 static Servo_HandleTypeDef servo_handles[SERVO_TIM_CHANNEL_MAX] = {0};
-static uint16_t              timer_arr = 0;  /* ARR 缓存，避免重复读 */
+static uint32_t              timer_arr = 0;  /* ARR 缓存，避免重复读 */
 
 /* ================================================================
  *  Servo_Init
@@ -30,9 +30,9 @@ void Servo_Init(TIM_HandleTypeDef *htim)
     timer_arr = htim->Init.Period;
     float pulse_per_us = (float)(timer_arr + 1)
                        / (SERVO_PWM_PERIOD_MS * 1000.0f);
-    uint16_t min_cmp = (uint16_t)(SERVO_PWM_MIN_US * pulse_per_us);
-    uint16_t max_cmp = (uint16_t)(SERVO_PWM_MAX_US * pulse_per_us);
-    uint16_t mid_cmp = (min_cmp + max_cmp) / 2;
+    uint32_t min_cmp = (uint32_t)(SERVO_PWM_MIN_US * pulse_per_us);
+    uint32_t max_cmp = (uint32_t)(SERVO_PWM_MAX_US * pulse_per_us);
+    uint32_t mid_cmp = (min_cmp + max_cmp) / 2;
 
     /* ---- 通道 2 = TIM_CHANNEL_3（TIM2→PB10 或 TIM5→PE8） ---- */
     uint8_t idx = 2;
@@ -109,7 +109,7 @@ void Servo_SetAngle(uint8_t ch, float angle)
     float    scale = (float)(servo_handles[ch].max_pulse
                            - servo_handles[ch].min_pulse)
                    / SERVO_ANGLE_RANGE;
-    uint16_t pulse = (uint16_t)(servo_handles[ch].min_pulse
+    uint32_t pulse = (uint32_t)(servo_handles[ch].min_pulse
                                 + angle * scale);
 
     __HAL_TIM_SET_COMPARE(servo_handles[ch].htim,
@@ -153,7 +153,7 @@ float Servo_GetAngle(uint8_t ch)
     if (ch >= SERVO_TIM_CHANNEL_MAX)          return -1.0f;
     if (!servo_handles[ch].initialized)       return -1.0f;
 
-    uint16_t pulse = __HAL_TIM_GET_COMPARE(servo_handles[ch].htim,
+    uint32_t pulse = __HAL_TIM_GET_COMPARE(servo_handles[ch].htim,
                                             servo_handles[ch].channel);
     float scale = (float)(servo_handles[ch].max_pulse
                         - servo_handles[ch].min_pulse)
